it now renders top and side view for circular
I’m not sure what happened but some how the call to the listener had
been deleted, as well as the private variables required for the
arguments for this method. all have been replaced and it works again.
there are still issues with the error handling which have been
commented out so that it runs.