fixed setText() for Outputs and general formatting
Discovered why setText() problem was happening and replaced a try/catch
block with an if to fix.
formatted the output string.
put units on GUI.
improved layout so TF’s are more uniform in size.
added functionality to check box so units and lbl’s now gray out as well