rendering illustrated elliptical orbits first attempt
I have completed this up to the point of just drawing the orbit. It
doesn’t work, but this might be because I haven’t completed
implementing the other features. this could be an issue because they
are all tied in together. Or something else isn’t working, haven’t had
time to test…