Added elliptical side view
Note - I stupidly added the ability for making it switch if arg go peri
was 180 but of course for anything in-between it makes no sence, the
point is to see it side on not end on which would be the case for a arg
of peri equaling 90 or 270.  I have commented out the code for now just
incase I change my mind and make it so it is a 3D rendering and hence
would want an end on or whatever view.