changed the main layout
The main layout was inefficient, I have made the main toolbar smaller
by making it the same width as the input are. This allows the output
area to be taller and hence the tab area to be squarer making the
rendering (no matter the RAAN angle) more even is size.