changed the method of storing the curve for the inclination angle
I had it so every shape in _inclination was being printed. this meant
that if the user changed inclination there were 2 shapes and both were
printed. I tried making it so it always added to index 0 and erasing
index 0 evertime the method was called but neither worked. The solution
was to only print index 0, since this is where I have set the val to go
every time. I could have just got the last val each time but didn’t see
much difference.